requirements.txt
requests
transformers
openai
attrs
pydantic
pymongo[srv]
rich